+++ api/v1/views/places_reviews.py
@@ -6,18 +6,21 @@ from models.review import Review
 from flask import jsonify, request, redirect
 from api.v1.views import app_views
 
+
 @app_views.route("/places/<place_id>/reviews", methods=['GET'])
 def reviews(place_id):
     """
     Returnns all reviews with the specified
     place_id
     """
-    obj = storage.all(cls="Review").values()
-    reviews = [item.to_dict() for item in obj if item.to_dict['place_id'] == place_id]
+    obj = storage.all(Review).values()
+    reviews = [item.to_dict() for item in obj 
+               if item.to_dict['place_id'] == place_id]
 
     if len(review) < 1:
-        return(redirect("/api/v1/nop"))
+        return(abort(404))
     return (jsonify(reviews))
+
 
 @app_views.route("/reviews/<review_id>", methods=['GET'])
 def review(review_id):
@@ -26,7 +29,7 @@ def review(review_id):
     """
     review = storage.get(Review, review_id)
     if review is None:
-        return (redirect("/api/v1/nop"))
+        return (abort(404))
     return (jsonify(review.to_dict()))
 
 @app_views.route("/reviews/<review_id>", methods=['DELETE'])
@@ -36,8 +39,11 @@ def delete_review(review_id):
     """
     review = storage.get(Place, review_id)
     if review is None:
-        return (redirect("/api/v1/nop"))
-    storage.delete(review)
+        return (abort(404))
+    for k, v in storage.all(Review).items():
+        if v == review:
+            storage.delete(review)
+            storage.save()
     return({}, 200)
 
 @app_views.route("/places/<place_id>/reviews", methods=['POST'])
@@ -45,20 +51,21 @@ def create_review(place_id):
     """
     Creates a new review
     """
-    if request.is_json():
+    if request.is_json:
         new_review = request.get_json()
         place = storage.get(Place, place_id)
         if place is None:
-            return (redirect("/api/v1/nop"))
+            return (abort(404))
         if 'user_id' not in new_review.keys():
             return (jsonify({"error": "Missing user_id"}), 400)
         user = storage.get(User, new_place['user_id'])
         if user is None:
-            return (redirect("/api/v1/nop"))
+            return (abort(404))
         if 'text' not in new_place.keys():
             return (jsonify({"error": "Missing text"}), 400)
-
-        model = Review(new_review)
+        
+        new_review["place_id"] = place_id
+        model = Review(**new_review)
         model.save()
         return (jsonify(model), 201)
     return (jsonify({"error": "Not a JSON"}), 400)
@@ -68,13 +75,13 @@ def update_review(review_id):
     """
     Updates the review
     """
-    if request.is_json():
+    if request.is_json:
         update_review = request.get_json()
         review = storage.get(Review, review_id)
         if review is None:
-            return (redirect("/api/v1/nop"))
+            return (abort(404))
 
-        setattr(review, update_review['name'])
+        setattr(review, "name", update_review['name'])
         review.save()
         return (jsonify(review.to_dict()), 200)
     return (jsonify({"error": "Not a JSON"}), 400)
